pyproject.toml
[project]
name = "gelab-zero-mcp-server"
version = "1.0.0"
description = "GELab-Zero MCP Server - 控制 Android 设备的 GUI Agent"
authors = [{name = "GELab Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "mcp",
    "pydantic",
    "yaml",
    "megfile",
    "Pillow",
    "opencv-python",
]

[project.scripts]
gelab-mcp-server = "mcp_server.standard_mcp_server:main"

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
